requirements/common.txt
django==5.1.7
django-extensions==3.2.3
django-environ==0.12.0
django-template-partials==24.4
django-htmx==1.22.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django_components==0.130
pillow==11.3.0
markdown==3.8.2
django-summernote==0.8.20.0